Moved classes to files Removed tokeninfo
