+++ WithingsToGarminSync/Models/General/GarminSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace WithingsToGarminSync.Models.General
+{
+	public class GarminSettings
+	{
+		public string Username { get; set; }
+		public string Password { get; set; }
+	}
+}
