+++ WithingsToGarminSync/Models/General/RunData.cs
@@ -1,0 +1,12 @@
+﻿using WithingsToGarminSync.Models.Withings;
+
+namespace WithingsToGarminSync.Models.General
+{
+	public class RunData
+	{
+		public WithingsAccessTokenBody Token { get; set; }
+		public DateTime? LastRun { get; set; }
+		public double LastWeight { get; set; }
+		public DateTime? LastWeightDate { get; set; }
+	}
+}
