+++ WithingsToGarminSync/Models/General/WithingsSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace WithingsToGarminSync.Models.General
+{
+	public class WithingsSettings
+	{
+		public string ClientId { get; set; }
+		public string ClientSecret { get; set; }
+		public string RedirectUrl { get; set; }
+	}
+}
